feat(container): convert to typescript
